r4ng3r-V2-main/rtf/requirements.txt
# RTF v2.0 Requirements
# Core
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
flask>=3.0.0
pydantic>=2.0.0
pyyaml>=6.0.1
rich>=13.7.0
httpx>=0.27.0
requests>=2.31.0

# DB / ORM
# (SQLite via stdlib — no extra dep)

# AWS
boto3>=1.34.0

# Azure
msal>=1.28.0

# GCP
google-auth>=2.29.0
google-api-python-client>=2.127.0

# LDAP
ldap3>=2.9.1

# Shodan
shodan>=1.31.0

# Reporting
jinja2>=3.1.3
markdown>=3.6
openpyxl>=3.1.2
reportlab>=4.1.0
weasyprint>=62.0; sys_platform != 'win32'

# Cryptography / Network
cryptography>=42.0.0
dnspython>=2.6.0

# Async
asyncio-throttle>=1.0.2

# Dev / Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
mypy>=1.9.0

# Extras (optional, for SOCMINT modules)
# pillow>=10.0.0
# python-docx>=1.1.0
